Compilation and documentation improvements (#86)
* Add documentation comment for vanilla djb2 hash algorithm

* Fix contradictory function arguments for `demo_state()`

* Make `rand_no_lint` static

Resolves some compiler warnings

* Add documentation comments for `loop.h` functions

* Fix compiler warning from `demo_state()`

- Conflicting argument names for `max_pieces` (used correct spelling)
- `const` declarations in header file serve no purpose

* Declare internal function `make_table__` as static

* Fix compiler warning about `make_table` not having a prototype

* Change return type of `message_len` from `int` to `size_t`

Result of this function is used for size comparisons and `memcpy`, hence it
makes sense for it to return a `size_t`

+++ src/loop.c
@@ -154,14 +154,14 @@ int main(int argc, char *argv[]) {
 }
 
 void read_message(int file_descriptor, int epoll_fd, client_state *state) {
-  int message_len = full_message_availiable(file_descriptor);
+  size_t message_len = full_message_availiable(file_descriptor);
 
   if (message_len) {
     uint8_t message[MAX_SIZE_MESSAGE_INT * 4];
     uint8_t message_type = 0;
     // https://pubs.opengroup.org/onlinepubs/007904975/functions/recv.html
     // Peek the message at the socket.
-    recv(file_descriptor, message, (size_t) message_len, 0);
+    recv(file_descriptor, message, message_len, 0);
     memcpy(&message_type, message + 4, 1); // NOLINT
      
     printf("processing message of type %d\n", (int)message_type);
